Support `copy` option in chatops_deployer.yml
`copy` is an array of strings in the format "<source>:<destination>"
source is the path to a file relative to /etc/chatops_deployer/copy in the deployer
destination is the path relative to chatops_deployer.yml to which the source file
should be copied. Copying of files happen soon after the repository is cloned
and before any docker containers are created.

Why is this needed?
This is one way of supporting environment specific configs.
`docker-compose` can read environment variables stored inside a file. So
if a developer is using a file, say `.env` for loading environment variables,
In staging environment, chatops_deployer will be able to overwrite .env
from a file that is present inside the chatops_deployer server stored at
a known location, say `/etc/chatops_deployer/copy/myapp/staging/.env`.
This way, secrets can be put inside the deployer out of band, and use
them later for deployments.